// src/app/models/data/producer-result.ts
export interface ProducerResultJson {
  id: number;
  producerGuid: string;
  producerType: string;
  firstName: string;
  lastName: string;
  npn: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  serviceState: string;
  subType: string;
  status: string;
  termDate: string;
  producerOrgId: string;
  producerOrgName: string;
  producerOrgGuid: string;
}